Fixed the real issue where padding would go outside of box
By general convention, when a box is defined, the padding is extra on the outside and get's added to the box rather than keeping the box the same size and adding the padding on the inside. By adding "border-box" attribute, the issue is fixed. 

The reason this is better to simply removing width attribute is in cases where not enough text is present for a box resulting in smaller box sizes than what is expected.

This was seen on the News page where a few listings had very little text.